Generate a Highcharts pie chart.

This function generates a Highcharts pie chart based on the given data.

hgch_pie(data, dic = NULL, var_cat = NULL, var_num = NULL, ...)

Arguments

data

A data frame containing the data to be plotted.

dic

A named list of character vectors that replaces column names in data. The names of the list should match column names in data, and each vector should contain the replacement names.

var_cat

A character vector of categorical variable(s) to use in the chart.

var_num

A character vector of numeric variable(s) to use in the chart.

...

Additional arguments to be passed to plot_opts.

Value

A Highcharts pie chart object.

See also

Examples

data1 <- data.frame(name = c("A", "B", "C"), y = c(10, 20, 30))
hgch_pie(data1, var_cat = "name", var_num = "y")
#> in theme_datasketch